Update todo.js
add the feature of re-asking and fixes some bug.

+++ todo.js
@@ -1,7 +1,7 @@
 let relevent;
-let arr = []
 let id;
 let next_id;
+const { count } = require('node:console');
 const fs = require('node:fs');
 // console.log("Enter relevent no. ")
 const readline = require('node:readline');
@@ -9,7 +9,8 @@ const rl = readline.createInterface({
     input: process.stdin,
     output: process.stdout
 });
-rl.question(` ======= Welcome to our todo app ========
+function option() {
+    rl.question(` ======= Welcome to our todo app ========
     1. Create todo
     2. Update todo
     3. Read todo
@@ -17,43 +18,55 @@ rl.question(` ======= Welcome to our todo app ========
     5. Search todo
     6. Quit
     Enter relevent no. `, relevent => {
-    console.log(relevent)
-    if (relevent == 6) {
-        console.log("program exit!")
-        rl.close()
-    }
-    else if (relevent == 1) {
-        rl.question(`Enter todo: `, todo => {
-            createTodo(todo)
+        console.log(relevent)
+        if (relevent == 6) {
+            console.log("program exit!")
             rl.close()
-        })
-    } else if (relevent == 2) {
-        rl.question(`Enter id which you want to update: `, id => {
+        }
+        else if (relevent == 1) {
             rl.question(`Enter todo: `, todo => {
-                updateTodo(id, todo)
-                rl.close()
+                createTodo(todo)
             })
-        })
-    } else if (relevent == 3) {
-        readTodo()
-        rl.close()
-    } else if (relevent == 4) {
-        rl.question(`id? `, id => {
-            deleteTodo(id)
-            rl.close()
-        })
-    } else if (relevent == 5) {
-        rl.question(`Id? `, id => {
-            searchTodo(id)
-            rl.close()
-        })
-    } else {
-        console.log("Program is exiting...")
-        rl.close()
-    }
-})
+        } else if (relevent == 2) {
+            if (!fs.existsSync('todo.json')) {
+                console.log("file doesn't exist first Enter 1 for create it.")
+                option()
+            } else {
+                rl.question(`Enter id which you want to update: `, id => {
+                    rl.question(`Enter todo: `, todo => {
+                        updateTodo(id, todo)
+                    })
+                })
+            }
+        } else if (relevent == 3) {
+            readTodo()
+        } else if (relevent == 4) {
+            if (!fs.existsSync('todo.json')) {
+                console.log("file doesn't exist first Enter 1 for create it.")
+                option()
+            } else {
+                rl.question(`id? `, id => {
+                    deleteTodo(id)
+                })
+            }
+        } else if (relevent == 5) {
+            if (!fs.existsSync('todo.json')) {
+                console.log("file doesn't exist first Enter 1 for create it.")
+                option()
+            } else {
+                rl.question(`Id? `, id => {
+                    searchTodo(id)
+                })
+            }
+        } else {
+            console.log("Invalid input! try again...")
+            option()
+        }
+    })
+}
 
 function createTodo(todo) {
+    let arr = []
     if (!fs.existsSync('todo.json')) {
         let obj = {
             "id": 1,
@@ -64,6 +77,7 @@ function createTodo(todo) {
         fs.writeFile("todo.json", done, err => {
             if (err) throw err;
             console.log("todo saved");
+            option()
         });
         console.log("todo created ")
     } else {
@@ -71,12 +85,19 @@ function createTodo(todo) {
             if (err) {
                 console.error(err)
             }
-            // console.log(data)
-            if (data.length > 0) {
+            // console.log(data.length)
+            if (data.length - 2 == 0){
+                next_id = 1
+                let obj = {
+                    "id": next_id,
+                    "todo": todo
+                }
+            }
+             else if (data.length > 0) {
                 new_data = JSON.parse(data)
                 // console.log(new_data)
-                for (let i = 0; i < new_data.length; i++) {
-                    arr.push(new_data[i])
+                for (let item of new_data) {
+                    arr.push(item)
                 }
                 next_id = new_data[new_data.length - 1]["id"] + 1;
             } else if (data.length == 0) {
@@ -92,6 +113,7 @@ function createTodo(todo) {
             fs.writeFile("todo.json", done, err => {
                 if (err) throw err;
                 console.log("todo updated");
+                option()
             });
         }
         )
@@ -100,6 +122,7 @@ function createTodo(todo) {
 
 
 function updateTodo(id, todo) {
+    let arr = []
     let counter = 0
     fs.readFile('todo.json', 'utf8', (err, data) => {
         if (err) {
@@ -112,46 +135,53 @@ function updateTodo(id, todo) {
             for (let i = 0; i < new_data.length; i++) {
                 if (id == new_data[i]["id"]) {
                     new_data[i]["todo"] = todo;
-                    counter += 1;
+                    counter++
                 }
                 arr.push(new_data[i])
             }
             if (counter == 0) {
-                console.log("Id not found")
-            } else {
-                console.log("updated successfully")
+                console.log("id not found")
+                option()
+            }
+            else {
+                console.log("todo updated")
+                option()
             }
         }
         // console.log(arr)
         let done = JSON.stringify(arr)
         fs.writeFile("todo.json", done, err => {
             if (err) throw err;
-            // console.log("todo updated");
         });
     }
     )
 }
 
 function readTodo() {
-    let data;
-    fs.readFile('todo.json', 'utf8', (err, data) => {
-        if (err) {
-            console.error(err)
-        }
-        // console.log(data)
-        if (data.length > 0) {
-            new_data = JSON.parse(data)
-            // console.log(new_data)
-            for (let i = 0; i < new_data.length; i++) {
-                console.log(`Id: ${new_data[i]["id"]} , Todo: ${new_data[i]["todo"]}`)
+    if (!fs.existsSync('todo.json')) {
+        console.log("file doesn't exist first Enter 1 for create it.")
+        option()
+    } else {
+        let arr = []
+        fs.readFile('todo.json', 'utf8', (err, data) => {
+            if (err) {
+                console.error(err)
             }
-
-        }
-    })
-    console.log("todo read")
+            // console.log(data)
+            if (data.length > 0) {
+                new_data = JSON.parse(data)
+                // console.log(new_data)
+                for (let i = 0; i < new_data.length; i++) {
+                    console.log(`Id: ${new_data[i]["id"]} , Todo: ${new_data[i]["todo"]}`)
+                }
+                option()
+            }
+        })
+    }
 }
 
 function deleteTodo(id) {
+    let arr = []
     let counter = 0
     fs.readFile('todo.json', 'utf8', (err, data) => {
         if (err) {
@@ -164,12 +194,21 @@ function deleteTodo(id) {
             for (let i = 0; i < new_data.length; i++) {
                 if (id != new_data[i]["id"]) {
                     arr.push(new_data[i])
-                    counter += 1
+                } else if (id == new_data[i]["id"]) {
+                    counter++
                 }
             }
         }
         // console.log(arr)
         let done = JSON.stringify(arr)
+        if (counter == 0) {
+            console.log("id not found")
+            option()
+        } else {
+            console.log("Deleted")
+            option()
+        }
+
         fs.writeFile("todo.json", done, err => {
             if (err) throw err;
             // console.log("todo updated");
@@ -177,16 +216,12 @@ function deleteTodo(id) {
     }
     )
     // console.log("todo updated")
-    if (counter == 0) {
-        console.log("Id not found")
-    } else {
-        console.log("deleted successfully")
-    }
 }
 
 function searchTodo(id) {
-    // console.log(id)
+    let arr = []
     let counter = 0
+    // console.log(id)
     fs.readFile('todo.json', 'utf8', (err, data) => {
         if (err) {
             console.error(err)
@@ -194,16 +229,21 @@ function searchTodo(id) {
         // console.log(data)
         else if (data.length > 0) {
             new_data = JSON.parse(data)
-            // console.log(new_data)
-            for (let i = 0; i < new_data.length; i++) {
-                if (id == new_data[i]["id"]) {
-                    console.log(`At the id ${id} the todo is ${new_data[i]["todo"]}`)
-                    counter += 1
+            // console.log(id)
+            for (let item of new_data) {
+                if (id == item.id) {
+                    console.log(`At the id ${id} the todo is ${item.todo}`)
+                    counter++
                 }
             }
             if (counter == 0) {
-                console.log("Id not found ")
+                console.log("id not found")
+                option()
+            } else {
+                option()
             }
         }
     })
 }
+
+option()
